+++ routes/categories.mjs
@@ -7,6 +7,9 @@ import {
   CAN_LIST_CATEGORY,
   CAN_REMOVE_CATEGORY,
 } from '../models/rights';
+import { csvToJson } from '../middlewares/csv-to-json';
+import contentNegociation from '../middlewares/content-negociation';
+import maskOutput from '../middlewares/mask-output';
 
 const router = generateCRUD(Category, {
   create: {
@@ -18,6 +21,8 @@ const router = generateCRUD(Category, {
       campus: 'campus._id',
     },
     middlewares: [
+      contentNegociation,
+      maskOutput,
       async (ctx, next) => {
         const searchParams = ctx.filters;
         if (ctx.query && ctx.query.search) {
@@ -43,6 +48,13 @@ const router = generateCRUD(Category, {
   },
   update: {
     right: CAN_EDIT_CATEGORY,
+  },
+  batch: {
+    right: CAN_CREATE_CATEGORY,
+    refs: ['_id', 'label'],
+    middlewares: [
+      csvToJson,
+    ],
   },
 });