+++ routes/ratings.mjs
@@ -4,13 +4,20 @@ import {
   CAN_CREATE_RATING,
   CAN_LIST_RATING,
 } from '../models/rights';
+import contentNegociation from '../middlewares/content-negociation';
+import maskOutput from '../middlewares/mask-output';
 
 const router = generateCRUD(Rating, {
   create: {
     right: CAN_CREATE_RATING,
+    successCode: 204,
   },
   list: {
     right: CAN_LIST_RATING,
+    middlewares: [
+      contentNegociation,
+      maskOutput,
+    ],
     async main(ctx) {
       const { offset, limit } = ctx.parseRangePagination(Rating);
       const [total, data] = await Promise.all([